requirements.txt
Flask==0.10.1
PyYAML==3.11
coverage==3.7.1
flake8==2.1.0
isoweek==1.3.0
nose==1.3.1
